Fix clang-tidy warnings for sidelink code

+++ examples/nr-v2x-examples/cttc-nr-v2x-demo-simple.cc
@@ -832,7 +832,7 @@ main(int argc, char* argv[])
     if (!useIPv6)
     {
         // Set Tx traces
-        for (uint16_t ac = 0; ac < clientApps.GetN(); ac++)
+        for (uint32_t ac = 0; ac < clientApps.GetN(); ac++)
         {
             Ipv4Address localAddrs = clientApps.Get(ac)
                                          ->GetNode()
@@ -849,7 +849,7 @@ main(int argc, char* argv[])
         }
 
         // Set Rx traces
-        for (uint16_t ac = 0; ac < serverApps.GetN(); ac++)
+        for (uint32_t ac = 0; ac < serverApps.GetN(); ac++)
         {
             Ipv4Address localAddrs = serverApps.Get(ac)
                                          ->GetNode()
@@ -868,7 +868,7 @@ main(int argc, char* argv[])
     else
     {
         // Set Tx traces
-        for (uint16_t ac = 0; ac < clientApps.GetN(); ac++)
+        for (uint32_t ac = 0; ac < clientApps.GetN(); ac++)
         {
             clientApps.Get(ac)->GetNode()->GetObject<Ipv6L3Protocol>()->AddMulticastAddress(
                 groupAddress6);
@@ -887,7 +887,7 @@ main(int argc, char* argv[])
         }
 
         // Set Rx traces
-        for (uint16_t ac = 0; ac < serverApps.GetN(); ac++)
+        for (uint32_t ac = 0; ac < serverApps.GetN(); ac++)
         {
             serverApps.Get(ac)->GetNode()->GetObject<Ipv6L3Protocol>()->AddMulticastAddress(
                 groupAddress6);
